requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
python-multipart==0.0.5
Pillow==8.4.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.9.1+cpu 
torchvision==0.10.1+cpu 
